requirements.txt
fastapi[standard]~=0.115.7
uvicorn[standard]~=0.34.0
bcrypt~=4.2.1
pyjwt~=1.7.1
requests~=2.32.3
python-dotenv~=1.0.1
passlib~=1.7.4
sqlalchemy~=2.0.37
sqlmodel~=0.0.22
pydantic~=2.10.6
starlette~=0.45.3